move BridgeServiceProcessor and corresponding impl and test to processor package
