+++ app/src/main/java/com/x8bit/bitwarden/data/platform/processor/BridgeServiceProcessor.kt
@@ -1,4 +1,4 @@
-package com.x8bit.bitwarden.data.platform.manager
+package com.x8bit.bitwarden.data.platform.processor
 
 import com.bitwarden.bridge.IBridgeService
